First cut at updates for the new Org and ESS.
For legibility, I've suppressed some whitespace differences in document.html in this commit.

I'm expecting to do away with the old code for encapsulating R evaluation (`ksource`) in favor of ESS's own encapsulation features—if I indeed get around to making that work. At this point, I'm only maintaining the R side of Daylight to help with my work on revisions to old papers, like Rickrack.